Dependencies: update requirement `pyyaml~=5.4` (#5060)
Earlier versions have critical security flaws that have been fixed in
`pyyaml==5.4`. Note that `plumpy` also needs to be upgraded to `0.20.0`
which adds support for this version of `pyyaml`.

The `UnsafeLoader` is replaced by the `Loader` which are identical, but
the former is only being kept as an alias for backwards compatibility
but it might be removed in future releases.

Cherry-pick: c78e0e27a60b6859dfd2ba66e0ded55ca2651ce4