CLI: show daemon worker load even when under threshold (#4663)
The extra query performed in `verdi process list` to get the load on the
daemon runners is slow - and usually results in no output for the user
in case that the load is below the threshold. With this change, the load
is shown even when below the threshold. Furthermore, the query is only
performed if the daemon is running and skipped otherwise.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>