Dependencies: pin `sqlalchemy` to minor version 1.3
The version `1.4` breaks our interface, most likely since we are
accessing private variables. This causes installs to fail since we were
not pinning the minor version. This problem was already addressed on the
`aiida-core==1.6` series, but that dropped support for Python 3.6. So
installs on that version would install `aiida-core==1.5` where the limit
was not yet added.