CI: switch from `postgresql-10` to the default
Since we use `ubuntu-lastest` as the image, which has now changed on
Github Actions and the current available version of PostgreSQL is no
longer 10 as we request. Since we don't really require a specific
version, we just leave it unspecified as to not have this break the next
time the ubuntu image changes.