Logging: add the `report` method and `REPORT` level to `logging` module
The built-in `logging` module comes with a number of pre-set logging
levels, such as `DEBUG` and `INFO`. AiiDA adds the `REPORT` level, which
sits in between `INFO` and `WARNING` and is designed to add some extra
granularity of control for messages. The report level is intended to be
used for messages that are purely informational, and so not a warning,
yet they should probably be shown by default. Traditionally, `INFO` log
messages logged by module code are often a bit too verbose to always
be showing by default. That is why AiiDA sets the default loglevel to
`REPORT` allowing code, such as `WorkChain`s to emit informational
messages that are shown by default, without also forcing all `INFO`
messages to be shown, which are typically too verbose.

To make the `REPORT` level behave like a built-in level, we hot-fix the
`logging.Logger` class to also have a `report` method, such that it can
be called as one would call `debug()` or `info()`.