CLI: add `-v/--verbosity` option to all `verdi` commands
The custom class used for the `verdi` command, `MostSimilarCommandGroup`
is renamed to `VerdiCommandGroup`. The reason for this is that it adds
new functionality, in addition to the existing functionality of
suggesting command names in case the command is not found, to
automatically add a `-v/--verbosity` option to all commands. This option
provides a choice to set the log level for the AiiDA logger. This allows
a user to control the verbosity of output that is emitted through the
logging system in a way that is consistent across all `verdi` commands.

Since through this method, the verbosity option is only added to sub
commands, the top-level commands `verdi` itself won't inherit from this
mechanism. Instead, the option is added manually and explicitly. Note
that in this case the callback of the verbosity, which sets the log
level on the profile, does not work, since for the top-level command the
verbosity option is parsed before the profile is set on the context.
That is why the processing of the verbosity is done explicitly in the
command itself.

By making the `-v` option available for all commands, the user can pass
it at only level of the sub command stack. There is one side-effect
however. Since `verdi` itself already had the `-v/--version` option,
which clashes with the `-v/--verbosity`, the `-v` shorthand for the
version option had to be removed. For consistency the `-h` shorthand for
the help flag, which is the only other option that is also available
across all commands, is also removed.

Note that this change also fixes a bug that only made the command name
suggestion functionality available at the top level but not for the sub
commands. By overriding the ``group`` method to use the same class as
the top-level `verdi` command, all subcommands should recursively have
the same functionality.

There is one subtlety where if the `cls` is set explicitly by one of the
subcommands, that should be respected. It can't simply override it with
`VerdiCommandGroup` because the subclass may have overridden the
`get_command` as well with custom functionality. This puts the onus on
the developer of subcommands with custom classes to make sure it
inherits from `VerdiCommandGroup` if it is going to overwrite the
`get_command` method and make sure it calls the super.

Finally, the ``run_cli_command`` fixture, designed to help test CLI
commands, needed to be updated to manually apply the ``VERBOSITY``
option. In testing, the sub command is direclty fetched from the module
where it is defined, instead of getting it through the ``click`` API
starting from the base ``verdi`` command. This means the functionality
that applies the ``VERBOSITY`` option is not hit and so it has to be
called manually.