CLI: pipe `cmdline.utils.echo` functions through logger
The utility functions in the `aiida.cmdline.utils.echo` module are
designed to be used in CLI commands to print output to stdout or stderr
streams. This is now changed to instead log the message to the
`aiida.cmdline` logger. This logger was added in the previous commit.
The reason is that the verbosity of this logger can be controlled
through the `--verbosity` option that was added to all commands.

To match the available log levels, the `echo_report` method was added,
which, as the name suggests, logs messages at the `REPORT` level. The
`echo` method, which was used to simply print a message without a
particular prefix, logs also at the `REPORT` level, since that is the
default log level in AiiDA. This makes sense since when developers used
`echo` they probably want it to show by default, but it is not a warning
and so should not be shown when the log level is set to `WARNING`.

The `echo_highlight` method has been removed as it was essentially a
thin wrapper around functionality that is already provided by `echo`.
Instead of using the different `color` argument, one can use `echo`
using the `fg` argument which is more consistent with the interface from
`click`'s methods which are wrapped.

All remaining direct calls of `click.echo` and `click.secho` in the code
have been replaced by the relevant `cmdline.utils.echo` command to make
sure that they are properly piped through the logging system and respect
the `--verbosity` option. Note that a few instances remain, for example,
in the `verdi completioncommand` command, but this is intentional as
they should not be influenced by the `--verbosity` option.