CLI: replace `echo_info` by `echo_report`
In the previous commit, the `echo` utility functions were changed to go
through the logger system, instead of directly using `click.echo` to
print to stdout and stderr. Since the default log level is `REPORT` in
AiiDA, messages logged with `echo_info` will no longer be shown by
default, unless the log level is set to `INFO` or `DEBUG`. To return to
the previous behavior, all `echo_info` instances are replaced by the new
`echo_report`, unless the existing message really corresponds to info
that is a bit verbose and can be hidden by default.

Calls to `info` for the `MIGRATE_LOGGER`, `EXPORT_LOGGER` and the
`DELETE_LOGGER` are changed to `report`. The current implementation
using these loggers assumed that the default log level would be `info`
and so these messages should be shown by default. This choice was made
since this functionality is mostly called through the command line. With
the current change where the default is now `report` these messages
would no longer show up by default.