Engine: add `CalcJobImporter` class and associated entry point group
The `CalcJobImporter` class is added, which defines a single abstract
staticmethod `parse_remote_data`. The idea is that plugins can define an
importer for a `CalcJob` implementation and implement this method. The
method takes a `RemoteData` node that points to a path on the associated
computer that contains the input and output files of a calculation that
has been run outside of AiiDA, but by an executable that is normally run
with this particular `CalcJob`.

The `parse_remote_data` implementation should read the input files found
in the remote data and parse their content into the input nodes that
when used to launch the calculation job, would result in similar input
files. These inputs, including the `RemoteData` as the `remote_folder`
input, can then be used to run an instance of this particular `CalcJob`.
The engine will recognize the `remote_folder` input, signalling an
import job, and instead of running a normal job that creates the input
files on the remote before submitting it to the scheduler, it passes
straight to the retrieve step. This will retrieve the files from the
`RemoteData` as if it would have been created by the job itself. If a
parsers was defined in the inputs, the contents are parsed and the
returned output nodes are attached.

The `CalcJobImporter` can be loaded through its entry point name using
the `CalcJobImporterFactory`, just like the entry points of all other
entry point groups have their associated factory. As a shortcut, the
`CalcJob` class, provides the `get_importer` class method which will
attempt to load a `CalcJobImporter` class with the exact same entry
point. Alternatively, the caller can specify the desired entry point
name should it not correspond to that of the `CalcJob` class.

To test the functionality, a `CalcJobImporter` is implemented for the
`ArithmeticAddCalculation` class.