Remove superfluous _validate from Data node
The Data nodeclass overwrote the `_validate` function but
 (a) did no validation
 (b) included commented code that referenced a private git repository
 (c) returned `None` instead of `True`.

The outcommented code involved validation of the `source` attribute and
the referenced issue pointed out that the value of this attribute was
not officially standardized (potentially leading users to using it in
different ways).

Since there has not been validation of this field since 2015, it seems
safe to assume that it cannot be easily reintroduced at this point.