CLI: remove `VERBOSE` option and replace with `VERBOSITY`
The `VERBOSE` option is obsolete since all commands now automatically
have the `VERBOSITY` option exposed. The `VERBOSE` option was mostly
used to configure a logger (most notably the `DELETE_LOGGER`) but this
is no longer necessary. Since those loggers are children of the main
`AIIDA_LOGGER`, whose level is controlled through the `VERBOSITY`
option, the child logger will inherit its value.

Some commands used the `VERBOSE` option more as a boolean switch. To
reproduce this behavior, a utiltiy function `is_verbose` is added to the
`cmdline` module, which will return `True` if the loglevel of the
`CMDLINE_LOGGER` is equal or smaller than the `INFO` level and `False`
otherwise.

The `verdi archive` commands used an ad-hoc implementation of the
verbosity option. This has been removed and they now also rely on the
logging level that is set on the `AIIDA_LOGGER` through the global
`VERBOSITY` option.