Logging: remove `override_log_formatter*` utilities
The `override_log_formatter` and `override_log_formatter_context`
utilities of the `aiida.common.log` module were used in certain CLI
commands to temporarily change the formatter of the logger. This was
done because they used the logging system as the way to output messages
to the terminal, but the default logger prefixed messages with extensive
information such as timestamps and the name of the logger, where only
the message was required.

These utilities are no longer needed since this custom formatting for
the CLI is now done ensured by configuring the `CliFormatter` custom
formatter which takes care of only including the log level as prefix.