CLI: ensure all loggers have a custom formatter
So far, the utility functions of the `aiida.cmdline.utils.echo` module
have been rerouted through the `CMDLINE_LOGGER` which properly formats
the log message with just the message itself. However, the loggers of
any other modules will inherit the formatter of the parent `AIIDA_LOGGER`
which will prefix the message with verbose content such as the
timestamp, name of the logger and more, which is not what we want for
output generated by CLI commands.

The solution is to define a custom formatter, `CliFormatter`, which is
added to the new module `aiida.cmdline.utils.log`, and which is
configured by default for the `CMDLINE_LOGGER`. The `configure_logging`
function is updated to now take a keyword argument `cli`, which when set
to `True`, will configure this new formatter for all loggers. By calling
this method with `cli=True` in the callback of the `VERBOSITY` option,
all loggers should be correctly configured to only print the message
with the log level prefixed, when invoked through the CLI.

There is one peculiarity where the `load_backend_if_not_loaded`
decorator utility also needs to call the `configure_logging` function.
The reason is that this function loads the database backend, which
indirectly calls the `configure_logging` to add the database log
handler, but this does not pass `cli=True`. This means that the CLI
specific configuration is undone.