CLI: fix bug in the `CliFormatter` log utility
The `format` method assumed that the `prefix` attribute would always
exist, which is not always this case. This would result in an uncaught
exception when a message would get logged. This went unnoticed since the
utility was not individually tested. Additionally, messages with literal
percentage signs would also except because it would always be
interpolated with the `record.args` but if that was an empty tuple, a
`TypeError` would be raised.