Dependencies: remove `click-completion`
As of `click==8.0` the full set of functionality that `click-completion`
provided for our tab-completion is now shipped with `click` itself, so
we can remove this dependency. The main difference is that the string
that is required to activate it, is slightly different and it is shell
dependent. The documentation has been updated to correctly reflect this.
Since it is now shell dependent, the old `verdi completioncommand`,
which was already (unofficially) deprecated is no longer correct. To
keep it correct, one would have to try and detect the shell but this is
clearly outside of the scope of `aiida-core` so we simply remote it.