Add proposal for repository maintain functionality
It would in principle consist of a single verdi command with one
optional flag and one optional argument:

  > verdi repository maintain [--live] [--pass-down <STR>]

The base command will perform all the available and necessary
maintainance operations but will require to lock down the profile
in order to proceed (it will ask the user to confirm before).

Optionally, the `--live` option can be passed so that only the
maintainance operations that are safe to perform while actively
using AiiDA will be applied.

Finnally, the user can also `--pass-down` a backend specific string
that will be interpreted by the backend method for even further
customization of the process.