Docs: add topics section on repository usage
This adds a badly needed section on how to interact with the file
repository through the `Node` interface. So far there was not a single
place dedicated to showing all repository related methods.

Also expose the `Repository.get_object` method on the level of the
`NodeRepositoryMixin` such that it is available from the `Node`
interface.