`Node`: add the `copy_tree` method
This method allows to easily copy the entire repository tree, or a
subdirectory, of a `Node` to a place on the local file system. The
method simply forwards to the `aiida.repository.Repository` class, where
the functionality is actually implemented using the `Repository.walk`
method, recursively looping over all the files contained within the
node's repository.