`CalcJob`: allow directories in `local_copy_list`
Up till now, the `local_copy_list` could only be used to specify
individual files that should be copied. However, very often, one may
have an input node from whose repository an entire directory, or sub
directory should be copied. The only way to do this was to manually
iterate over the contents of that repository directory and add the files
to the `local_copy_list` one by one.

Here we extend the syntax of the `local_copy_list` and the second
argument can now also point to a directory in the repository of the
source node. Its entire contents will be copied to the relative path
defined by the third element in the tuple. Note that the directory
itself won't be copied, just its contents.