`Dict`: implement the `__eq__` method to compare plain dictionary
The `__eq__` method is overridden to not only let a `Dict` node compare
true to itself, but also to the plain dictionary that represents its
value. That is to say, the following:

    node = Dict(dict={'a': 1'})
    assert node == node.value

now passes the assert. This brings the behavior of `Dict` on par with
all the other base types, `Bool`, `Float`, `Int`, `Str` and `Bool`.