`List`: register the class with the `to_aiida_type` serializer
This will allow users to pass a plain list as input for a port that
accepts `List` nodes and specified the `to_aiida_type` dispatch as
serializer. The value will be automatically converted to a `List` node.
This brings the functionality for `List` on par with the other base type
data node classes.