`verdi process list`: fix logging bug in `check_worker_load`
The `check_worker_load` utility, called in `verdi process list`, caused
an exception in the logging because it logged a message including a
literal percentage sign. However, percentage signs carry special meaning
as they are placeholders for argument substitution and literal
percentage signs need to be escaped by doubling the character.