`BaseRestartWorkChain`: clearify the log report when output ports not compatible
The `BaseRestartWorkChain` output ports required are reported to missing
in two conditions.
1) when the nested `CalcJob` do not correctly giving the outputs.
2) when the base work chain does not correctly expose the output ports,
which is not considered here. It causes issue https://github.com/aiidateam/aiida-quantumespresso/issues/723
These shold be two different scenario but we report at one log.
 Would be better to separate this two situation but that will complicate the code logic here
(need to check the actual outputs of embed process and then raise for the first half report),
I don't think it worth to do for this.