📚 DOCS: Add example for `retrieve_temporary_list` (#5157)
The current section on how the user can retrieve files temporarily to use
during parsing doesn't contain any examples on how to use this feature. Here
we add an example on how to specify the list of files to be retrieved in a
temporary folder, clearly showing the `retrieve_temporary_list` attribute of
the `CalcInfo`. We also add an example of how to obtain the
`retrieved_temporary_folder` from the `kwargs` of the `Parser.parse()` method.