`Dict`: fix the `__eq__` implementation to call super
The `__eq__` method was recently implemented for the `Dict` data type
but it erroneously did not call through to the super implementation in
case the dictionary content does not match that of `other`. This caused
the fallback equality implemented on the `Node` base class, which
ensures that two nodes with the same UUID compare equal, to be missed.