`Data`: allow `source` to be passed as a keyword argument
The `Data` class defines the `source` property which allows to get and
set the `source` attribute. It was designed to hold information on the
source of the data if it came directly from an external database.

Prior to `aiida-core` v1.0 it was possible to set this attribute
directly through the constructor as a keyword argument, but due to the
refactoring of the ORM, this was no longer possible. It went undetected
because it was never tested and the two data types where this feature is
typically used, the `UpfData` and `CifData`, implemented the keyword
argument themselves, hiding the fact that `Data` no longer supported it.

The `Data` constructor signature is updated to once again recognize the
`source` keyword argument, and the custom implementation is removed from
the `UpfData` and `CifData` plugins which now simply rely on the super
class.

A section is added to the "How to work with Data" section in the docs to
document this feature and to recommend its usage when constructing data
nodes by hand from external databases.