Docker: Default to number of physical cores for localhost computer. (#5161)
The `configure-aiida.sh` script that is used to configure the instance
in the Docker image now allows to configure the `--mpiprocs-per-machine`
argument for the `verdi computer setup` command that sets up the host
machine as the `localhost` computer. The value to be set can be defined
through the `LOCALHOST_MPI_PROCS_PER_MACHINE` environment variable.

The default value for it is based on the actual number of physical cores
of the host machine which is determined by calling `psutil.cpu_count`
through a Python interpreter. It is necessary here to use the method of
the `psutil` library instead of that of the `os` standard module, since
the former allows to get just the physical cores and not also the logical
ones, by passing `logical=False` as an argument. Since `psutil` is an
explicit dependency of `aiida-core`, relying on its presence is fine.
If this automatic determination of the number of physical cores fails
and the environment variable has not been set, a warning is printed and
the script is exited.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>