CLI: Warn for already existing ``code@computer label`` in ``verdi code setup`` and ``duplicate``
Implement reprompting mechanism in ``verdi code setup`` for code label/computer combinations that are already in the database. The Computer option gains callback checking the DB for the set code label/computer combination

If it is found the user is asked if he would like to change the label. If that is the case the label is set to `None`, which triggers a second label argument, which will perform the same check as the computer option but it will raise an error and reprompt instead. This second label argument is ``hidden=True, expose_value=False`` to avoid duplicate help messages and name clashes with the other label option

Works both for remote and local codes. For non-interactive commands an error is thrown